+++ types/component-extensions.d.ts
@@ -6,15 +6,15 @@ import { CustomElementOptions } from '@ficusjs/core'
 
 export * from '@ficusjs/component-extensions'
 
-export declare class ExtensionBuilder<EB, AS, AP, EO> {
-  newInstance: ExtensionBuilder<EB, AS, AP, EO>
-  withBreakpointConfig (breakpointConfig: BreakpointConfig): ExtensionBuilder<EB, AS, AP, EO>
-  withEventBus (eventBus: EventBus<EB>): ExtensionBuilder<EB, AS, AP, EO>
-  withI18n (i18n: I18n): ExtensionBuilder<EB, AS, AP, EO>
-  withLazyRender (): ExtensionBuilder<EB, AS, AP, EO>
-  withStyles (): ExtensionBuilder<EB, AS, AP, EO>
-  withLocalState (): ExtensionBuilder<EB, AS, AP, EO>
-  withStore (store: AppStateStore<AS, AP>): ExtensionBuilder<EB, AS, AP, EO>
-  withWorkerStore (worker: Worker): ExtensionBuilder<EB, AS, AP, EO>
-  create (options: CustomElementOptions<EO>): CustomElementOptions<EO>
+export declare class ExtensionBuilder<EB, AS extends object, AP, EO, ET> {
+  newInstance: ExtensionBuilder<EB, AS, AP, EO, ET>
+  withBreakpointConfig (breakpointConfig: BreakpointConfig): ExtensionBuilder<EB, AS, AP, EO, ET>
+  withEventBus (eventBus: EventBus<EB>): ExtensionBuilder<EB, AS, AP, EO, ET>
+  withI18n (i18n: I18n): ExtensionBuilder<EB, AS, AP, EO, ET>
+  withLazyRender (): ExtensionBuilder<EB, AS, AP, EO, ET>
+  withStyles (): ExtensionBuilder<EB, AS, AP, EO, ET>
+  withLocalState (): ExtensionBuilder<EB, AS, AP, EO, ET>
+  withStore (store: AppStateStore<AS>): ExtensionBuilder<EB, AS, AP, EO, ET>
+  withWorkerStore (worker: Worker): ExtensionBuilder<EB, AS, AP, EO, ET>
+  create (options: CustomElementOptions<EO, ET>): CustomElementOptions<EO, ET>
 }