+++ types/component.d.ts
@@ -4,8 +4,8 @@ export interface FicusComponent<TS> extends HTMLElement {
   setState: (stateFn: (state: TS) => Partial<TS>, callback: () => void) => void
 }
 
-export type ComponentOptions<TS, TE> = CustomElementOptions<TE> & {
+export type ComponentOptions<TS, TE, TP> = CustomElementOptions<TE, TP> & {
   state?: () => TS
 }
 
-export declare function createComponent<TS, TE>(tagName: string, options: ComponentOptions<TS, TE>): void
+export declare function createComponent<TS, TE, TP>(tagName: string, options: ComponentOptions<TS, TE, TP>): void